Refactor the server routes so that each route is per-file rather than each resource is per-file
This makes the directory a lot easier to navigate. Note that the routing pattern has also changed such that we register routes using our own helper functions. This makes it easier if we were to potentially ever sort the objects ourselves going forward, for example. It is also similar to existing patterns used by alex-c-line to register its commands.
