+++ apps/back-end/src/server/routes/auth/helpers/createCallbackUrl.ts
@@ -1,0 +1,5 @@
+function createCallbackUrl(originalUrl: string = "/api/v1/auth/google/callback") {
+  return `${process.env.API_BASE_URL!}${originalUrl}`;
+}
+
+export default createCallbackUrl;
