+++ apps/back-end/src/server/routes/index.ts
@@ -2,12 +2,13 @@ import type { Express } from "express";
 
 import { CodeError } from "@alextheman/utility/v6";
 
-import authRouter from "src/server/routes/auth";
-import blogsRouter from "src/server/routes/blogs";
-import currentUserRouter from "src/server/routes/currentUser";
-import usersRouter from "src/server/routes/users";
+import initialiseAuthRouter from "src/server/routes/auth";
+import initialiseBlogsRouter from "src/server/routes/blogs";
+import initialiseCurrentUserRouter from "src/server/routes/currentUser";
+import initialiseUsersRouter from "src/server/routes/users";
 import defineEndpoint from "src/utility/defineEndpoint";
 import handleEndpointMiddleware from "src/utility/handleEndpointMiddleware";
+import initialiseRouter from "src/utility/initialiseRouter";
 import requireAuth from "src/utility/validators/requireAuth";
 
 function createEndpoints(app: Express) {
@@ -27,10 +28,10 @@ function createEndpoints(app: Express) {
     }),
   );
 
-  app.use(defineEndpoint("current-user"), currentUserRouter);
-  app.use(defineEndpoint("users"), usersRouter);
-  app.use(defineEndpoint("auth"), authRouter);
-  app.use(defineEndpoint("blogs"), blogsRouter);
+  initialiseRouter(app, "auth", initialiseAuthRouter);
+  initialiseRouter(app, "blogs", initialiseBlogsRouter);
+  initialiseRouter(app, "current-user", initialiseCurrentUserRouter);
+  initialiseRouter(app, "users", initialiseUsersRouter);
 
   app.get(defineEndpoint("control/be-error"), () => {
     throw new CodeError(
