+++ apps/back-end/src/server/routes/users/index.ts
@@ -1,0 +1,12 @@
+import type { Router } from "express";
+
+import getUserById from "src/server/routes/users/getUserById";
+import registerEndpoints from "src/utility/registerEndpoints";
+
+function initialiseUsersRouter(router: Router) {
+  registerEndpoints(router, {
+    getUserById,
+  });
+}
+
+export default initialiseUsersRouter;
